Uses new exception macro.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@881 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ geometry/geodetector/src/GeoMaterials.cc
@@ -182,7 +182,7 @@ TGeoElement* GeoMaterials::readElement(GearDir& elementContent, double& weight)
   //Collect information
   weight            = elementContent.getParamNumValue("attribute::weight");
   string elemName   = elementContent.getParamString("attribute::name");
-  double atomNumber = elementContent.getParamNumValue("AtomNumber");
+  int    atomNumber = static_cast<int>(elementContent.getParamNumValue("AtomNumber"));
   double massNumber = elementContent.getParamNumValue("MassNumber");
 
   //Build root element
